+++ apps/counter-dom-esm/.eslintrc.cjs
@@ -1,0 +1,2 @@
+const rootConfig = require("../../.eslintrc.cjs");
+module.exports = rootConfig;