+++ apps/counter-dom-js/.eslintrc.cjs
@@ -1,7 +1,0 @@
-const rootConfig = require("../../.eslintrc.cjs");
-module.exports = {
-  ...rootConfig,
-  parserOptions: {
-    ...rootConfig.parserOptions,
-  },
-};