+++ apps/counter-react-ts-edit-context/src/logic.ts
@@ -1,0 +1,9 @@
+import type { Immutable } from "@lauf/store";
+
+export interface CounterState {
+  counter: number;
+}
+
+export const INITIAL_STATE: Immutable<CounterState> = {
+  counter: 0,
+} as const;