+++ apps/counter-react-ts/vite.config.ts
@@ -1,13 +1,19 @@
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
 import legacy from "@vitejs/plugin-legacy";
+import react from "@vitejs/plugin-react";
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  build: {
+    minify: "terser",
+    rollupOptions: {
+      treeshake: true,
+    },
+  },
   plugins: [
     react(),
     legacy({
-      targets: ["defaults", "not IE 11"],
+      targets: ["defaults"],
     }),
   ],
   clearScreen: false,