+++ apps/counter-react-js/CHANGELOG.md
@@ -1,5 +1,91 @@
 # counter-react-js
 
+## 1.3.3-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@2.0.0-beta.1
+  - @lauf/store@2.0.0-beta.1
+
+## 1.3.3-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@2.0.0-beta.0
+  - @lauf/store@2.0.0-beta.0
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [a4c5ce2]
+- Updated dependencies [1e5ad2a]
+- Updated dependencies [44a2490]
+- Updated dependencies
+  - @lauf/store-react@1.3.2
+  - @lauf/store@1.3.2
+
+## 1.3.2-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.2-alpha.2
+  - @lauf/store@1.3.2-alpha.2
+
+## 1.3.2-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.2-alpha.1
+  - @lauf/store@1.3.2-alpha.1
+
+## 1.3.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.2-alpha.0
+  - @lauf/store@1.3.2-alpha.0
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [3b623c6]
+- Updated dependencies [547f408]
+- Updated dependencies
+- Updated dependencies [5df19f9]
+  - @lauf/store-react@1.3.1
+  - @lauf/store@1.3.1
+
+## 1.3.1-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.1-alpha.2
+  - @lauf/store@1.3.1-alpha.2
+
+## 1.3.1-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.1-alpha.1
+  - @lauf/store@1.3.1-alpha.1
+
+## 1.3.1-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @lauf/store-react@1.3.1-alpha.0
+  - @lauf/store@1.3.1-alpha.0
+
 ## 1.3.0
 
 ### Patch Changes